@@ -1,8 +1,9 @@
 package com.example.anujsharma.manageit;
 
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentManager;
-import android.support.v4.app.FragmentPagerAdapter;
+
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentManager;
+import androidx.fragment.app.FragmentPagerAdapter;
 
 public class MyViewPagerAdapter extends FragmentPagerAdapter {
 
