@@ -2,11 +2,12 @@ package com.example.anujsharma.manageit;
 
 import android.content.Context;
 import android.database.Cursor;
-import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
+
+import androidx.recyclerview.widget.RecyclerView;
 
 public class MyTaggedImagesViewAdapter extends RecyclerView.Adapter<MyTaggedImagesViewAdapter.MyImagesViewHolder> {
 
