@@ -1,7 +1,9 @@
 package com.example.anujsharma.manageit;
 
-import android.support.v7.widget.RecyclerView;
+
 import android.util.SparseBooleanArray;
+
+import androidx.recyclerview.widget.RecyclerView;
 
 import java.util.ArrayList;
 import java.util.List;
