@@ -3,10 +3,11 @@ package com.example.anujsharma.manageit;
 import android.content.Context;
 import android.net.Uri;
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
+import androidx.fragment.app.Fragment;
 
 public class TaggedImagesFragment extends Fragment {
 
